Adding the variable creation for attribute and constant Adding the texture parsing

+++ Lumina/include/lumina_metatokenizer.hpp
@@ -23,7 +23,7 @@ namespace Lumina
 
 		void expendInclude();
 		TypeDescriptor parseTypeDescriptor();
-		size_t parseArraySize();
+		int parseArraySize();
 		VariableDescriptor parseVariableDescriptor();
 		Condition parseCondition();
